RAG_ALL/5.27_chroma/API_Local_Embedding/requirement.txt
pdfminer
pdfminer.six
python-docx==1.1.2
openpyxl==3.1.3
dashscope==1.15.0
chromadb==0.5.0
zhipuai==2.1.0
torch==2.3.1
transformers==4.41.2
smodelscope
